modules/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
bleach==6.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.105.0
filelock==3.13.1
flake8==7.0.0
fonttools==4.47.0
fsspec==2023.12.2
idna==3.6
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.3
kaggle==1.6.3
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
openpyxl==3.1.2
openpyxl-image-loader==1.0.5
packaging==23.2
pandas==2.2.0
parso==0.8.3
Pillow==10.1.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.2.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-slugify==8.0.2
pytz==2023.4
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.27.0
svgwrite==1.4.3
sympy==1.12
text-unidecode==1.3
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
webencodings==0.5.1
